move findJUnitDependency
