+++ src/main/kotlin/org/jetbrains/research/testspark/actions/TestSparkAction.kt
@@ -97,7 +97,7 @@ class TestSparkAction : AnAction() {
 
         private val cardLayout = CardLayout()
 
-        private val llmSetupPanelFactory = LLMSetupPanelFactory()
+        private val llmSetupPanelFactory = LLMSetupPanelFactory(e)
         private val llmSampleSelectorFactory = LLMSampleSelectorFactory(project)
         private val evoSuitePanelFactory = EvoSuitePanelFactory()
 
@@ -108,15 +108,13 @@ class TestSparkAction : AnAction() {
 
                 val panel = JPanel(cardLayout)
 
-                val junit = findJUnitDependency(e)
-
                 panel.add(getMainPanel(), "1")
-                panel.add(createCardPanel(evoSuitePanelFactory, junit), "2")
-                panel.add(createCardPanel(llmSetupPanelFactory, junit), "3")
+                panel.add(createCardPanel(evoSuitePanelFactory), "2")
+                panel.add(createCardPanel(llmSetupPanelFactory), "3")
 
                 panel.add(
                     JBScrollPane(
-                        createCardPanel(llmSampleSelectorFactory, junit),
+                        createCardPanel(llmSampleSelectorFactory),
                         JBScrollPane.VERTICAL_SCROLLBAR_ALWAYS,
                         JBScrollPane.HORIZONTAL_SCROLLBAR_NEVER,
                     ),
@@ -145,33 +143,13 @@ class TestSparkAction : AnAction() {
             }
         }
 
-        private fun createCardPanel(toolPanelFactory: PanelFactory, junit: JUnitVersion?): JPanel {
+        private fun createCardPanel(toolPanelFactory: PanelFactory): JPanel {
             val cardPanel = JPanel(BorderLayout())
             cardPanel.add(toolPanelFactory.getTitlePanel(), BorderLayout.NORTH)
-            cardPanel.add(toolPanelFactory.getMiddlePanel(junit), BorderLayout.CENTER)
+            cardPanel.add(toolPanelFactory.getMiddlePanel(), BorderLayout.CENTER)
             cardPanel.add(toolPanelFactory.getBottomPanel(), BorderLayout.SOUTH)
 
             return cardPanel
-        }
-
-        private fun findJUnitDependency(e: AnActionEvent): JUnitVersion? {
-            val project = e.project!!
-            val virtualFile = e.getData(CommonDataKeys.VIRTUAL_FILE_ARRAY)?.firstOrNull() ?: return null
-
-            val index = ProjectRootManager.getInstance(project).fileIndex
-            val module = index.getModuleForFile(virtualFile) ?: return null
-
-            for (orderEntry in ModuleRootManager.getInstance(module).orderEntries) {
-                if (orderEntry is LibraryOrderEntry) {
-                    val libraryName = orderEntry.library?.name ?: continue
-                    for (junit in JUnitVersion.values()) {
-                        if (libraryName.contains(junit.groupId)) {
-                            return junit
-                        }
-                    }
-                }
-            }
-            return null
         }
 
         /**
