+++ src/main/kotlin/org/jetbrains/research/testspark/actions/llm/LLMSetupPanelFactory.kt
@@ -1,5 +1,10 @@
 package org.jetbrains.research.testspark.actions.llm
 
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.actionSystem.CommonDataKeys
+import com.intellij.openapi.roots.LibraryOrderEntry
+import com.intellij.openapi.roots.ModuleRootManager
+import com.intellij.openapi.roots.ProjectRootManager
 import com.intellij.openapi.ui.ComboBox
 import com.intellij.ui.components.JBLabel
 import com.intellij.util.ui.FormBuilder
@@ -17,7 +22,7 @@ import javax.swing.JLabel
 import javax.swing.JPanel
 import javax.swing.JTextField
 
-class LLMSetupPanelFactory : PanelFactory {
+class LLMSetupPanelFactory(private val e: AnActionEvent) : PanelFactory {
     private val defaultModulesArray = arrayOf("")
     private var modelSelector = ComboBox(defaultModulesArray)
     private var llmUserTokenField = JTextField(30)
@@ -61,10 +66,10 @@ class LLMSetupPanelFactory : PanelFactory {
      * and a user token field. These components are stylized using the `stylizeMainComponents` method.
      * The UI labels for the platform, token, and model components are retrieved using the
      * `TestSpark*/
-    override fun getMiddlePanel(junit: JUnitVersion?): JPanel {
+    override fun getMiddlePanel(): JPanel {
         stylizeMainComponents(platformSelector, modelSelector, llmUserTokenField, llmPlatforms, settingsState)
 
-        junitSelector.detected = junit
+        junitSelector.detected = findJUnitDependency()
 
         return FormBuilder.createFormBuilder()
             .setFormLeftIndent(10)
@@ -93,6 +98,26 @@ class LLMSetupPanelFactory : PanelFactory {
                 false,
             )
             .panel
+    }
+
+    private fun findJUnitDependency(): JUnitVersion? {
+        val project = e.project!!
+        val virtualFile = e.getData(CommonDataKeys.VIRTUAL_FILE_ARRAY)?.firstOrNull() ?: return null
+
+        val index = ProjectRootManager.getInstance(project).fileIndex
+        val module = index.getModuleForFile(virtualFile) ?: return null
+
+        for (orderEntry in ModuleRootManager.getInstance(module).orderEntries) {
+            if (orderEntry is LibraryOrderEntry) {
+                val libraryName = orderEntry.library?.name ?: continue
+                for (junit in JUnitVersion.values()) {
+                    if (libraryName.contains(junit.groupId)) {
+                        return junit
+                    }
+                }
+            }
+        }
+        return null
     }
 
     /**
