+++ src/main/kotlin/org/jetbrains/research/testspark/actions/template/PanelFactory.kt
@@ -1,6 +1,5 @@
 package org.jetbrains.research.testspark.actions.template
 
-import org.jetbrains.research.testspark.data.JUnitVersion
 import org.jetbrains.research.testspark.services.SettingsApplicationService
 import org.jetbrains.research.testspark.settings.SettingsApplicationState
 import javax.swing.JButton
@@ -12,7 +11,7 @@ interface PanelFactory {
 
     fun getTitlePanel(): JPanel
 
-    fun getMiddlePanel(junit: JUnitVersion?): JPanel
+    fun getMiddlePanel(): JPanel
 
     fun getBottomPanel(): JPanel
